% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VosonComments.R
\docType{data}
\name{VosonComments}
\alias{VosonComments}
\title{vosonComments}
\format{
A data frame with 24 rows and 12 columns:
\describe{
  \item{Comment}{Content of the comment}
  \item{AuthorDisplayName}{Displayed name of the comment author}
  \item{AuthorProfileImageUrl}{URL of the comment author's profile picture}
  \item{AuthorChannelUrl}{URL of the comment author's channel}
  \item{AuthorChannelID}{Unique ID of the comment author's channel}
  \item{ReplyCount}{Number of replies to the comment}
  \item{LikeCount}{Number of likes on the comment}
  \item{PublishedAt}{Date and time the comment was published}
  \item{UpdatedAt}{Date and time the comment was last updated}
  \item{CommentID}{Unique ID of the comment}
  \item{ParentID}{Unique ID of the comment to which the comment is a reply}
  \item{VideoID}{Unique ID of the video to which the comment is a reply}
}
}
\source{
The data was modeled to resemble a random subset of 24 comments from the following YouTube video:
\url{https://www.youtube.com/watch?v=1aheRpmurAo}, exported with the tuber package. It was manually recreated to comply with
YouTube's terms of service while still preserving the general structure for demonstration purposes.
}
\usage{
VosonComments
}
\description{
An example dataset to showcase the structure of YouTube comments extracted
using the \code{\link[vosonSML:Collect]{Collect}}
 function from the \pkg{vosonSML} package.
}
\keyword{datasets}
